I made a few changes to the the footer and 'google apps' image of my google hompage and also made a few edits to the google search result page
On the previous code, the google apps icon on the google homepage was not in line with the anchor links (i.e Gmail and Images) and the 'sign in' button, it was that way cos i had little experience on how to align everything. So, now i made use of made use of the <position: relative;> and <top, left, rigt, and bottom> functions to fix the icon in the right position.
As for the footer, I used the <border-bottom> function on the previous code to demarcate <div class=footer1> from the <p>  which didnt let the line cut across the width of the page after I moved the <div class=footer1> by 50px from the left. So, i corrected that by applying the <Hr> function immediately after the <div class=footer1> on my html code.

Moving on to the google search result page, I didn't make much changes to it, I just set the <div id=top> to <position: fixed;>
I set it to fixed so that the user can easily use the google search box even after he/she has scrolled down the page, instead of scrolling back up the page to acces the search box.